# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
      "ActivityProgressDTO": {
        "properties": {
          "best_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Duration Ms"
          },
          "completed": {
            "title": "Completed",
            "type": "boolean"
          },
          "completion_count": {
            "title": "Completion Count",
            "type": "integer"
          }
        },
        "required": [
          "completed",
          "completion_count",
          "best_duration_ms"
        ],
        "title": "ActivityProgressDTO",
        "type": "object"
      },
...
      "GenerationDraftRequestDTO": {
        "description": "Request to draft (or refine) a pack from a topic.\n\n``history`` is the opaque JSON message history returned by a prior draft\nturn (see :class:`GenerationDraftResponseDTO`); replaying it lets a\nrefinement turn keep the model's context. It is ``None`` on the first turn.",
        "properties": {
          "history": {
            "anyOf": [
              {
                "items": {},
                "maxItems": 200,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "History"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "topic": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "topic"
        ],
        "title": "GenerationDraftRequestDTO",
        "type": "object"
      },
      "GenerationDraftResponseDTO": {
        "description": "A drafted pack plus the updated conversation history to persist.\n\nThe client holds ``history`` between turns and passes it back on the next\n:class:`GenerationDraftRequestDTO` to refine the draft.",
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/PackDraft"
          },
          "history": {
            "items": {},
            "title": "History",
            "type": "array"
          }
        },
        "required": [
          "draft",
          "history"
        ],
        "title": "GenerationDraftResponseDTO",
        "type": "object"
      },
      "GenerationSavePackRequestDTO": {
        "description": "Request to persist a finalized :class:`PackDraft` as an owned pack.",
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/PackDraft"
          }
        },
        "required": [
          "draft"
        ],
        "title": "GenerationSavePackRequestDTO",
        "type": "object"
      },
...
      "PackCharacterDTO": {
        "properties": {
          "hanzi": {
            "title": "Hanzi",
            "type": "string"
          },
          "meaning": {
            "title": "Meaning",
            "type": "string"
          },
          "pinyin": {
            "title": "Pinyin",
            "type": "string"
          }
        },
        "required": [
          "hanzi",
          "pinyin",
          "meaning"
        ],
        "title": "PackCharacterDTO",
        "type": "object"
      },
      "PackDetailDTO": {
        "properties": {
          "char_count": {
            "title": "Char Count",
            "type": "integer"
          },
          "characters": {
            "items": {
              "$ref": "#/components/schemas/PackCharacterDTO"
            },
            "title": "Characters",
            "type": "array"
          },
          "color": {
            "title": "Color",
            "type": "string"
          },
          "glyph": {
            "title": "Glyph",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "owned": {
            "title": "Owned",
            "type": "boolean"
          },
          "progress": {
            "$ref": "#/components/schemas/PackProgressDTO"
          },
          "sentence_count": {
            "title": "Sentence Count",
            "type": "integer"
          },
          "sentences": {
            "items": {
              "$ref": "#/components/schemas/PackSentenceDTO"
            },
            "title": "Sentences",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "glyph",
          "color",
          "char_count",
          "sentence_count",
          "owned",
          "progress",
          "characters",
          "sentences"
        ],
        "title": "PackDetailDTO",
        "type": "object"
      },
      "PackDraft": {
        "description": "Structured output the generation agent returns for a practice pack.",
        "properties": {
          "characters": {
            "items": {
              "$ref": "#/components/schemas/PackDraftCharacter"
            },
            "maxItems": 30,
            "minItems": 1,
            "title": "Characters",
            "type": "array"
          },
          "coverage_note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Note"
          },
          "sentences": {
            "items": {
              "$ref": "#/components/schemas/PackDraftSentence"
            },
            "maxItems": 12,
            "title": "Sentences",
            "type": "array"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "characters"
        ],
        "title": "PackDraft",
        "type": "object"
      },
      "PackDraftCharacter": {
        "description": "A single drafted character with its model-generated gloss.",
        "properties": {
          "hanzi": {
            "maxLength": 1,
            "minLength": 1,
            "title": "Hanzi",
            "type": "string"
          },
          "meaning": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Meaning",
            "type": "string"
          },
          "pinyin": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Pinyin",
            "type": "string"
          }
        },
        "required": [
          "hanzi",
          "pinyin",
          "meaning"
        ],
        "title": "PackDraftCharacter",
        "type": "object"
      },
      "PackDraftSentence": {
        "description": "A drafted example sentence for the sentence-tracing activity.",
        "properties": {
          "hanzi": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Hanzi",
            "type": "string"
          },
          "pinyin": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Pinyin",
            "type": "string"
          },
          "translation": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Translation",
            "type": "string"
          }
        },
        "required": [
          "hanzi",
          "pinyin",
          "translation"
        ],
        "title": "PackDraftSentence",
        "type": "object"
      },
      "PackProgressDTO": {
        "properties": {
          "match": {
            "$ref": "#/components/schemas/ActivityProgressDTO"
          },
          "sentence": {
            "$ref": "#/components/schemas/ActivityProgressDTO"
          },
          "trace": {
            "$ref": "#/components/schemas/ActivityProgressDTO"
          }
        },
        "required": [
          "trace",
          "match",
          "sentence"
        ],
        "title": "PackProgressDTO",
        "type": "object"
      },
...
      "PackSentenceDTO": {
        "properties": {
          "hanzi": {
            "title": "Hanzi",
            "type": "string"
          },
          "pinyin": {
            "title": "Pinyin",
            "type": "string"
          },
          "translation": {
            "title": "Translation",
            "type": "string"
          }
        },
        "required": [
          "hanzi",
          "pinyin",
          "translation"
        ],
        "title": "PackSentenceDTO",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Learn to write Chinese characters by tracing them, stroke by stroke.",
    "title": "habagou",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/generation/draft": {
      "post": {
        "description": "Draft (or refine) a corpus-grounded pack for the caller's topic.",
        "operationId": "generate_draft_api_v1_generation_draft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationDraftRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationDraftResponseDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Model selection requires an admin account"
          },
          "422": {
            "description": "Requested model is not selectable"
          },
          "429": {
            "description": "Per-user generation rate limit exceeded"
          },
          "502": {
            "description": "Pack generation failed"
          },
          "503": {
            "description": "Pack generation is not configured"
          }
        },
        "summary": "Generate Draft",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/generation/packs": {
      "post": {
        "description": "Persist a finalized draft as a pack owned by the caller.",
        "operationId": "save_pack_api_v1_generation_packs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationSavePackRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackDetailDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "Draft references non-corpus characters"
          }
        },
        "summary": "Save Pack",
        "tags": [
          "generation"
        ]
      }
    },
...
  }
}